refactor(plugins): move samples to plugin registry repo, decouple tests
Samples belong with other plugins, not in the SDK host's repo. They
now live at Kb2uka/openhpsdr-zeus-plugins/samples/ — see the
companion PR #1 on that repo. This commit removes them from Zeus and
rewrites the end-to-end integration tests so they don't depend on
the samples being co-located.

Removed:
- samples/plugins/HelloWorld/, samples/plugins/Amplifier/ (moved to
  Kb2uka/openhpsdr-zeus-plugins repo).
- Zeus.slnx /samples/ folder + the two sample csproj entries.
- tests/Zeus.Plugins.Host.Tests.csproj — the two sample ProjectReferences
  and the CopySamplePlugins MSBuild target are gone. The
  CopyVstBridgeDylib target stays (the native bridge is a Zeus host
  concern, not a plugin author concern).
- tests/Zeus.Plugins.Host.Tests/SamplePluginTests.cs (replaced).

Added:
- tests/Zeus.Plugins.Host.Tests/EndToEndPluginTests.cs — same three
  scenarios SamplePluginTests covered (HelloWorld load + manifest,
  IBackendPlugin endpoint round-trip via TestServer, /api/plugins
  listing) but built on RoslynFixture so no external plugin
  artifacts are required. Source for the Amplifier fixture is the
  minimum surface needed to exercise IBackendPlugin.MapEndpoints
  through the host's REST mount.

RoslynFixture (tests/Zeus.Plugins.Host.Tests/RoslynFixture.cs):
- Reference assembly discovery upgraded. Old behaviour: scan only
  AppDomain.GetAssemblies(), which misses anything lazy-loaded. New
  behaviour: combine GetAssemblies + every .dll under
  AppContext.BaseDirectory + every .dll under the
  Microsoft.AspNetCore.App shared framework dir (located via
  typeof(object).Assembly.Location). This lets IBackendPlugin
  fixtures (which reference Microsoft.AspNetCore.Routing /
  Microsoft.AspNetCore.Http) compile without the test project
  pulling the AspNetCore assemblies into bin/.

Docs:
- author-guide.md "Where to go next" now points at the registry
  repo's samples/ instead of the in-repo samples/ that no longer
  exist.

Build + tests (Zeus.slnx):
- 0 warnings, 0 errors
- 1133 tests pass across 7 projects (was 1133 before the move; tests
  count unchanged, just sourced differently — 3 sample integration
  tests now use Roslyn fixtures).
