+++ tests/Zeus.Plugins.Host.Tests/RoslynFixture.cs
@@ -40,18 +40,65 @@ internal sealed class RoslynFixture : IDisposable
             Guid.NewGuid().ToString("N"));
         Directory.CreateDirectory(tempRoot);
 
-        // Reference assemblies — pull from the test host's resolved
-        // assembly list so the synthetic plugin sees exactly the types
-        // the host runtime exposes. Filter out anything whose backing
-        // file no longer exists on disk (previous test fixture's
-        // compiled plugin lingers in AppDomain after we delete the
-        // temp dir during Dispose).
-        var refs = AppDomain.CurrentDomain.GetAssemblies()
-            .Where(a => !a.IsDynamic && !string.IsNullOrEmpty(a.Location))
-            .Where(a => File.Exists(a.Location))
-            .Select(a => MetadataReference.CreateFromFile(a.Location))
-            .Cast<MetadataReference>()
-            .ToList();
+        // Reference assemblies — combine two sources:
+        //   1. Currently-loaded assemblies (AppDomain.GetAssemblies) —
+        //      catches everything actively in use plus its transitive
+        //      load graph.
+        //   2. Every .dll next to the test binary that wasn't lazy-loaded
+        //      yet. AspNetCore.Routing / Builder are typical examples:
+        //      the test project references them but they don't load
+        //      until used, so GetAssemblies misses them. Filling from
+        //      the output dir picks them up unconditionally.
+        var byPath = new Dictionary<string, MetadataReference>(StringComparer.OrdinalIgnoreCase);
+
+        foreach (var a in AppDomain.CurrentDomain.GetAssemblies())
+        {
+            if (a.IsDynamic || string.IsNullOrEmpty(a.Location)) continue;
+            if (!File.Exists(a.Location)) continue;
+            byPath[a.Location] = MetadataReference.CreateFromFile(a.Location);
+        }
+
+        foreach (var dll in Directory.EnumerateFiles(AppContext.BaseDirectory, "*.dll"))
+        {
+            if (!byPath.ContainsKey(dll))
+            {
+                try { byPath[dll] = MetadataReference.CreateFromFile(dll); }
+                catch { /* non-managed dll, skip */ }
+            }
+        }
+
+        // The Microsoft.AspNetCore.App shared framework dlls live next
+        // to the runtime, not in bin/. Locate the AspNetCore runtime
+        // dir from typeof(object)'s installation root and add every
+        // .dll under it. Without this, fixtures that implement
+        // IBackendPlugin can't compile (Microsoft.AspNetCore.Routing
+        // isn't loaded by the test harness on its own).
+        var coreDir = Path.GetDirectoryName(typeof(object).Assembly.Location);
+        if (coreDir is not null)
+        {
+            // coreDir = .../shared/Microsoft.NETCore.App/<ver>
+            // shared dir is two levels up, then add Microsoft.AspNetCore.App.
+            var sharedRoot = Path.GetFullPath(Path.Combine(coreDir, "..", "..", "Microsoft.AspNetCore.App"));
+            if (Directory.Exists(sharedRoot))
+            {
+                var aspnetVersionDir = Directory.EnumerateDirectories(sharedRoot)
+                    .OrderByDescending(d => d)
+                    .FirstOrDefault();
+                if (aspnetVersionDir is not null)
+                {
+                    foreach (var dll in Directory.EnumerateFiles(aspnetVersionDir, "*.dll"))
+                    {
+                        if (!byPath.ContainsKey(dll))
+                        {
+                            try { byPath[dll] = MetadataReference.CreateFromFile(dll); }
+                            catch { /* skip */ }
+                        }
+                    }
+                }
+            }
+        }
+
+        var refs = byPath.Values.ToList();
 
         // Ensure Zeus.Plugins.Contracts is present even if not yet loaded.
         var contractsAsm = typeof(Zeus.Plugins.Contracts.IZeusPlugin).Assembly;
